Update addresses to match EIP
