+++ packages/evm/src/precompiles/10-bls12-g2msm.ts
@@ -13,14 +13,14 @@ import {
 import type { ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
 
-export async function precompile11(opts: PrecompileInput): Promise<ExecResult> {
+export async function precompile10(opts: PrecompileInput): Promise<ExecResult> {
   const mcl = (<any>opts._EVM)._mcl!
 
   const inputData = opts.data
 
   if (inputData.length === 0) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12G2MSM (0x0f) failed: Empty input`)
+      opts._debug(`BLS12G2MSM (0x10) failed: Empty input`)
     }
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_INPUT_EMPTY), opts.gasLimit) // follow Geths implementation
   }
@@ -45,7 +45,7 @@ export async function precompile11(opts: PrecompileInput): Promise<ExecResult> {
   const gasUsed = (gasUsedPerPair * BigInt(numPairs) * BigInt(gasDiscountMultiplier)) / BigInt(1000)
   if (opts._debug !== undefined) {
     opts._debug(
-      `Run BLS12G2MSM (0x0f) precompile data=${short(opts.data)} length=${
+      `Run BLS12G2MSM (0x10) precompile data=${short(opts.data)} length=${
         opts.data.length
       } gasLimit=${opts.gasLimit} gasUsed=${gasUsed}`
     )
@@ -53,14 +53,14 @@ export async function precompile11(opts: PrecompileInput): Promise<ExecResult> {
 
   if (opts.gasLimit < gasUsed) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12G2MSM (0x0f) failed: OOG`)
+      opts._debug(`BLS12G2MSM (0x10) failed: OOG`)
     }
     return OOGResult(opts.gasLimit)
   }
 
   if (inputData.length % 288 !== 0) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12G2MSM (0x0f) failed: Invalid input length length=${inputData.length}`)
+      opts._debug(`BLS12G2MSM (0x10) failed: Invalid input length length=${inputData.length}`)
     }
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_INVALID_INPUT_LENGTH), opts.gasLimit)
   }
@@ -88,7 +88,7 @@ export async function precompile11(opts: PrecompileInput): Promise<ExecResult> {
       )
       if (!(equalsBytes(slicedBuffer, zeroBytes16) === true)) {
         if (opts._debug !== undefined) {
-          opts._debug(`BLS12G2MSM (0x0f) failed: Point not on curve`)
+          opts._debug(`BLS12G2MSM (0x10) failed: Point not on curve`)
         }
         return EvmErrorResult(new EvmError(ERROR.BLS_12_381_POINT_NOT_ON_CURVE), opts.gasLimit)
       }
@@ -98,7 +98,7 @@ export async function precompile11(opts: PrecompileInput): Promise<ExecResult> {
       G2 = BLS12_381_ToG2Point(opts.data.subarray(pairStart, pairStart + 256), mcl)
     } catch (e: any) {
       if (opts._debug !== undefined) {
-        opts._debug(`BLS12G2MSM (0x0f) failed: ${e.message}`)
+        opts._debug(`BLS12G2MSM (0x10) failed: ${e.message}`)
       }
       return EvmErrorResult(e, opts.gasLimit)
     }
@@ -113,7 +113,7 @@ export async function precompile11(opts: PrecompileInput): Promise<ExecResult> {
   const returnValue = BLS12_381_FromG2Point(result)
 
   if (opts._debug !== undefined) {
-    opts._debug(`BLS12G2MSM (0x0f) return value=${bytesToHex(returnValue)}`)
+    opts._debug(`BLS12G2MSM (0x10) return value=${bytesToHex(returnValue)}`)
   }
 
   return {
