+++ packages/evm/src/precompiles/11-bls12-pairing.ts
@@ -11,7 +11,7 @@ import type { PrecompileInput } from './types.js'
 const zeroBuffer = new Uint8Array(32)
 const oneBuffer = concatBytes(new Uint8Array(31), hexToBytes('0x01'))
 
-export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
+export async function precompile11(opts: PrecompileInput): Promise<ExecResult> {
   const mcl = (<any>opts._EVM)._mcl!
 
   const inputData = opts.data
@@ -20,7 +20,7 @@ export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
 
   if (inputData.length === 0) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12PAIRING (0x10) failed: Empty input`)
+      opts._debug(`BLS12PAIRING (0x11) failed: Empty input`)
     }
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_INPUT_EMPTY), opts.gasLimit)
   }
@@ -31,7 +31,7 @@ export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
   const gasUsed = baseGas + gasUsedPerPair * BigInt(Math.floor(inputData.length / 384))
   if (opts._debug !== undefined) {
     opts._debug(
-      `Run BLS12PAIRING (0x10) precompile data=${short(opts.data)} length=${
+      `Run BLS12PAIRING (0x11) precompile data=${short(opts.data)} length=${
         opts.data.length
       } gasLimit=${opts.gasLimit} gasUsed=${gasUsed}`
     )
@@ -39,14 +39,14 @@ export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
 
   if (inputData.length % 384 !== 0) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12PAIRING (0x10) failed: Invalid input length length=${inputData.length}`)
+      opts._debug(`BLS12PAIRING (0x11) failed: Invalid input length length=${inputData.length}`)
     }
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_INVALID_INPUT_LENGTH), opts.gasLimit)
   }
 
   if (opts.gasLimit < gasUsed) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12PAIRING (0x10) failed: OOG`)
+      opts._debug(`BLS12PAIRING (0x11) failed: OOG`)
     }
     return OOGResult(opts.gasLimit)
   }
@@ -75,7 +75,7 @@ export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
       )
       if (!equalsBytes(slicedBuffer, zeroBytes16)) {
         if (opts._debug !== undefined) {
-          opts._debug(`BLS12PAIRING (0x10) failed: Point not on curve`)
+          opts._debug(`BLS12PAIRING (0x11) failed: Point not on curve`)
         }
         return EvmErrorResult(new EvmError(ERROR.BLS_12_381_POINT_NOT_ON_CURVE), opts.gasLimit)
       }
@@ -85,7 +85,7 @@ export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
       G1 = BLS12_381_ToG1Point(opts.data.subarray(pairStart, pairStart + 128), mcl)
     } catch (e: any) {
       if (opts._debug !== undefined) {
-        opts._debug(`BLS12PAIRING (0x10) failed: ${e.message}`)
+        opts._debug(`BLS12PAIRING (0x11) failed: ${e.message}`)
       }
       return EvmErrorResult(e, opts.gasLimit)
     }
@@ -96,7 +96,7 @@ export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
       G2 = BLS12_381_ToG2Point(opts.data.subarray(g2start, g2start + 256), mcl)
     } catch (e: any) {
       if (opts._debug !== undefined) {
-        opts._debug(`BLS12PAIRING (0x10) failed: ${e.message}`)
+        opts._debug(`BLS12PAIRING (0x11) failed: ${e.message}`)
       }
       return EvmErrorResult(e, opts.gasLimit)
     }
@@ -132,7 +132,7 @@ export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
   }
 
   if (opts._debug !== undefined) {
-    opts._debug(`BLS12PAIRING (0x10) return value=${bytesToHex(returnValue)}`)
+    opts._debug(`BLS12PAIRING (0x11) return value=${bytesToHex(returnValue)}`)
   }
 
   return {
