+++ packages/evm/src/precompiles/12-bls12-map-fp-to-g1.ts
@@ -8,7 +8,7 @@ import { BLS12_381_FromG1Point, BLS12_381_ToFpPoint } from './util/bls12_381.js'
 import type { ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
 
-export async function precompile13(opts: PrecompileInput): Promise<ExecResult> {
+export async function precompile12(opts: PrecompileInput): Promise<ExecResult> {
   const mcl = (<any>opts._EVM)._mcl!
 
   const inputData = opts.data
@@ -17,7 +17,7 @@ export async function precompile13(opts: PrecompileInput): Promise<ExecResult> {
   const gasUsed = opts.common.paramByEIP('gasPrices', 'Bls12381MapG1Gas', 2537) ?? BigInt(0)
   if (opts._debug !== undefined) {
     opts._debug(
-      `Run BLS12MAPFPTOG1 (0x11) precompile data=${short(opts.data)} length=${
+      `Run BLS12MAPFPTOG1 (0x12) precompile data=${short(opts.data)} length=${
         opts.data.length
       } gasLimit=${opts.gasLimit} gasUsed=${gasUsed}`
     )
@@ -25,14 +25,14 @@ export async function precompile13(opts: PrecompileInput): Promise<ExecResult> {
 
   if (opts.gasLimit < gasUsed) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12MAPFPTOG1 (0x11) failed: OOG`)
+      opts._debug(`BLS12MAPFPTOG1 (0x12) failed: OOG`)
     }
     return OOGResult(opts.gasLimit)
   }
 
   if (inputData.length !== 64) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12MAPFPTOG1 (0x11) failed: Invalid input length length=${inputData.length}`)
+      opts._debug(`BLS12MAPFPTOG1 (0x12) failed: Invalid input length length=${inputData.length}`)
     }
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_INVALID_INPUT_LENGTH), opts.gasLimit)
   }
@@ -41,7 +41,7 @@ export async function precompile13(opts: PrecompileInput): Promise<ExecResult> {
   const zeroBytes16 = new Uint8Array(16)
   if (!equalsBytes(opts.data.subarray(0, 16), zeroBytes16)) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12MAPFPTOG1 (0x11) failed: Point not on curve`)
+      opts._debug(`BLS12MAPFPTOG1 (0x12) failed: Point not on curve`)
     }
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_POINT_NOT_ON_CURVE), opts.gasLimit)
   }
@@ -53,7 +53,7 @@ export async function precompile13(opts: PrecompileInput): Promise<ExecResult> {
     Fp1Point = BLS12_381_ToFpPoint(opts.data.subarray(0, 64), mcl)
   } catch (e: any) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12MAPFPTOG1 (0x11) failed: ${e.message}`)
+      opts._debug(`BLS12MAPFPTOG1 (0x12) failed: ${e.message}`)
     }
     return EvmErrorResult(e, opts.gasLimit)
   }
@@ -64,7 +64,7 @@ export async function precompile13(opts: PrecompileInput): Promise<ExecResult> {
   const returnValue = BLS12_381_FromG1Point(result)
 
   if (opts._debug !== undefined) {
-    opts._debug(`BLS12MAPFPTOG1 (0x11) return value=${bytesToHex(returnValue)}`)
+    opts._debug(`BLS12MAPFPTOG1 (0x12) return value=${bytesToHex(returnValue)}`)
   }
 
   return {
