+++ packages/evm/src/precompiles/13-bls12-map-fp2-to-g2.ts
@@ -3,21 +3,21 @@ import { bytesToHex, equalsBytes, short } from '@ethereumjs/util'
 import { EvmErrorResult, OOGResult } from '../evm.js'
 import { ERROR, EvmError } from '../exceptions.js'
 
-import { BLS12_381_FromG1Point, BLS12_381_ToG1Point } from './util/bls12_381.js'
+import { BLS12_381_FromG2Point, BLS12_381_ToFp2Point } from './util/bls12_381.js'
 
 import type { ExecResult } from '../types.js'
 import type { PrecompileInput } from './types.js'
 
-export async function precompile0c(opts: PrecompileInput): Promise<ExecResult> {
+export async function precompile13(opts: PrecompileInput): Promise<ExecResult> {
   const mcl = (<any>opts._EVM)._mcl!
 
   const inputData = opts.data
 
   // note: the gas used is constant; even if the input is incorrect.
-  const gasUsed = opts.common.paramByEIP('gasPrices', 'Bls12381G1AddGas', 2537) ?? BigInt(0)
+  const gasUsed = opts.common.paramByEIP('gasPrices', 'Bls12381MapG2Gas', 2537) ?? BigInt(0)
   if (opts._debug !== undefined) {
     opts._debug(
-      `Run BLS12G1ADD (0x0a) precompile data=${short(opts.data)} length=${
+      `Run BLS12MAPFP2TOG2 (0x13) precompile data=${short(opts.data)} length=${
         opts.data.length
       } gasLimit=${opts.gasLimit} gasUsed=${gasUsed}`
     )
@@ -25,14 +25,14 @@ export async function precompile0c(opts: PrecompileInput): Promise<ExecResult> {
 
   if (opts.gasLimit < gasUsed) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12G1ADD (0x0a) failed: OOG`)
+      opts._debug(`BLS12MAPFP2TOG2 (0x13) failed: OOG`)
     }
     return OOGResult(opts.gasLimit)
   }
 
-  if (inputData.length !== 256) {
+  if (inputData.length !== 128) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12G1ADD (0x0a) failed: Invalid input length length=${inputData.length}`)
+      opts._debug(`BLS12MAPFP2TOG2 (0x13) failed: Invalid input length length=${inputData.length}`)
     }
     return EvmErrorResult(new EvmError(ERROR.BLS_12_381_INVALID_INPUT_LENGTH), opts.gasLimit)
   }
@@ -42,39 +42,36 @@ export async function precompile0c(opts: PrecompileInput): Promise<ExecResult> {
   const zeroByteCheck = [
     [0, 16],
     [64, 80],
-    [128, 144],
-    [192, 208],
   ]
 
   for (const index in zeroByteCheck) {
     const slicedBuffer = opts.data.subarray(zeroByteCheck[index][0], zeroByteCheck[index][1])
     if (!(equalsBytes(slicedBuffer, zeroBytes16) === true)) {
       if (opts._debug !== undefined) {
-        opts._debug(`BLS12G1ADD (0x0a) failed: Point not on curve`)
+        opts._debug(`BLS12MAPFP2TOG2 (0x13) failed: Point not on curve`)
       }
       return EvmErrorResult(new EvmError(ERROR.BLS_12_381_POINT_NOT_ON_CURVE), opts.gasLimit)
     }
   }
 
-  // convert input to mcl G1 points, add them, and convert the output to a Uint8Array.
-  let mclPoint1
-  let mclPoint2
+  // convert input to mcl Fp2 point
+
+  let Fp2Point
   try {
-    mclPoint1 = BLS12_381_ToG1Point(opts.data.subarray(0, 128), mcl)
-    mclPoint2 = BLS12_381_ToG1Point(opts.data.subarray(128, 256), mcl)
+    Fp2Point = BLS12_381_ToFp2Point(opts.data.subarray(0, 64), opts.data.subarray(64, 128), mcl)
   } catch (e: any) {
     if (opts._debug !== undefined) {
-      opts._debug(`BLS12G1ADD (0x0a) failed: ${e.message}`)
+      opts._debug(`BLS12MAPFP2TOG2 (0x13) failed: ${e.message}`)
     }
     return EvmErrorResult(e, opts.gasLimit)
   }
+  // map it to G2
+  const result = Fp2Point.mapToG2()
 
-  const result = mcl.add(mclPoint1, mclPoint2)
-
-  const returnValue = BLS12_381_FromG1Point(result)
+  const returnValue = BLS12_381_FromG2Point(result)
 
   if (opts._debug !== undefined) {
-    opts._debug(`BLS12G1ADD (0x0a) return value=${bytesToHex(returnValue)}`)
+    opts._debug(`BLS12MAPFP2TOG2 (0x13) return value=${bytesToHex(returnValue)}`)
   }
 
   return {
