+++ packages/evm/src/precompiles/index.ts
@@ -11,15 +11,15 @@ import { precompile07 } from './07-ecmul.js'
 import { precompile08 } from './08-ecpairing.js'
 import { precompile09 } from './09-blake2f.js'
 import { precompile0a } from './0a-kzg-point-evaluation.js'
-import { precompile0c } from './0c-bls12-g1add.js'
-import { precompile0d } from './0d-bls12-g1mul.js'
-import { precompile0e } from './0e-bls12-g1msm.js'
-import { precompile0f } from './0f-bls12-g2add.js'
-import { precompile10 } from './10-bls12-g2mul.js'
-import { precompile11 } from './11-bls12-g2msm.js'
-import { precompile12 } from './12-bls12-pairing.js'
-import { precompile13 } from './13-bls12-map-fp-to-g1.js'
-import { precompile14 } from './14-bls12-map-fp2-to-g2.js'
+import { precompile0b } from './0b-bls12-g1add.js'
+import { precompile0c } from './0c-bls12-g1mul.js'
+import { precompile0d } from './0d-bls12-g1msm.js'
+import { precompile0e } from './0e-bls12-g2add.js'
+import { precompile0f } from './0f-bls12-g2mul.js'
+import { precompile10 } from './10-bls12-g2msm.js'
+import { precompile11 } from './11-bls12-pairing.js'
+import { precompile12 } from './12-bls12-map-fp-to-g1.js'
+import { precompile13 } from './13-bls12-map-fp2-to-g2.js'
 
 import type { PrecompileFunc, PrecompileInput } from './types.js'
 import type { Common } from '@ethereumjs/common'
@@ -147,7 +147,15 @@ const precompileEntries: PrecompileEntry[] = [
     precompile: precompile0a,
     name: 'KZG (0x0a)',
   },
-  // 0x00..0b: beacon block root, see PR 2810
+  {
+    address: '000000000000000000000000000000000000000b',
+    check: {
+      type: PrecompileAvailabilityCheck.EIP,
+      param: 2537,
+    },
+    precompile: precompile0b,
+    name: 'BLS12_G1ADD',
+  },
   {
     address: '000000000000000000000000000000000000000c',
     check: {
@@ -155,7 +163,7 @@ const precompileEntries: PrecompileEntry[] = [
       param: 2537,
     },
     precompile: precompile0c,
-    name: 'BLS12_G1ADD',
+    name: 'BLS12_G1MUL',
   },
   {
     address: '000000000000000000000000000000000000000d',
@@ -164,7 +172,7 @@ const precompileEntries: PrecompileEntry[] = [
       param: 2537,
     },
     precompile: precompile0d,
-    name: 'BLS12_G1MUL',
+    name: 'BLS12_G1MSM',
   },
   {
     address: '000000000000000000000000000000000000000e',
@@ -173,7 +181,7 @@ const precompileEntries: PrecompileEntry[] = [
       param: 2537,
     },
     precompile: precompile0e,
-    name: 'BLS12_G1MSM',
+    name: 'BLS12_G2ADD',
   },
   {
     address: '000000000000000000000000000000000000000f',
@@ -182,7 +190,7 @@ const precompileEntries: PrecompileEntry[] = [
       param: 2537,
     },
     precompile: precompile0f,
-    name: 'BLS12_G2ADD',
+    name: 'BLS12_G2MUL',
   },
   {
     address: '0000000000000000000000000000000000000010',
@@ -191,7 +199,7 @@ const precompileEntries: PrecompileEntry[] = [
       param: 2537,
     },
     precompile: precompile10,
-    name: 'BLS12_G2MUL',
+    name: 'BLS12_G2MSM',
   },
   {
     address: '0000000000000000000000000000000000000011',
@@ -200,7 +208,7 @@ const precompileEntries: PrecompileEntry[] = [
       param: 2537,
     },
     precompile: precompile11,
-    name: 'BLS12_G2MSM',
+    name: 'BLS12_PAIRING',
   },
   {
     address: '0000000000000000000000000000000000000012',
@@ -209,7 +217,7 @@ const precompileEntries: PrecompileEntry[] = [
       param: 2537,
     },
     precompile: precompile12,
-    name: 'BLS12_PAIRING',
+    name: 'BLS12_MAP_FP_TO_G1',
   },
   {
     address: '0000000000000000000000000000000000000013',
@@ -218,15 +226,6 @@ const precompileEntries: PrecompileEntry[] = [
       param: 2537,
     },
     precompile: precompile13,
-    name: 'BLS12_MAP_FP_TO_G1',
-  },
-  {
-    address: '0000000000000000000000000000000000000014',
-    check: {
-      type: PrecompileAvailabilityCheck.EIP,
-      param: 2537,
-    },
-    precompile: precompile14,
     name: 'BLS12_MAP_FP2_TO_G2',
   },
 ]
@@ -242,7 +241,7 @@ const precompiles: Precompiles = {
   '0000000000000000000000000000000000000008': precompile08,
   '0000000000000000000000000000000000000009': precompile09,
   '000000000000000000000000000000000000000a': precompile0a,
-  // 0b: beacon block root see PR 2810
+  '000000000000000000000000000000000000000b': precompile0b,
   '000000000000000000000000000000000000000c': precompile0c,
   '000000000000000000000000000000000000000d': precompile0d,
   '000000000000000000000000000000000000000e': precompile0e,
@@ -251,7 +250,6 @@ const precompiles: Precompiles = {
   '0000000000000000000000000000000000000011': precompile11,
   '0000000000000000000000000000000000000012': precompile12,
   '0000000000000000000000000000000000000013': precompile13,
-  '0000000000000000000000000000000000000014': precompile14,
 }
 
 type DeletePrecompile = {
